Extend KPM metrics (per-UE, per-nexran slice).
Refactor indexing scheme; PDCP data now correct.  Add PRB metrics to
the mac/scheduler and report them.  There are issues -- first, srslte
metrics are not indexed by cell, so we're not correct for > 1 cell.
Second, some metrics (e.g. the MAC) are zeroed every time they are
polled, so the MAC metrics only work for a single subscription and/or
if user doesn't toggle metrics reports on console.